requirements.txt
langchain-core 
langgraph
langchain
langchain-groq
fastapi
uvicorn
python-dotenv
pydantic